DB-9421 Add configuration to disable Spark accumulators (#3472)
* DB-9421 Add configuration to disable Spark accumulators

* DB-9421 Fix ITs

* DB-9421 Address spotbugs issues

* DB-9421 More SpotBugs
