+++ hbase_sql/src/main/java/com/splicemachine/derby/stream/spark/SparkDataSetProcessor.java
@@ -16,6 +16,7 @@ package com.splicemachine.derby.stream.spark;
 
 import com.google.common.base.Joiner;
 import com.google.common.collect.Lists;
+import com.splicemachine.EngineDriver;
 import com.splicemachine.access.HConfiguration;
 import com.splicemachine.access.api.FileInfo;
 import com.splicemachine.db.iapi.error.StandardException;
@@ -76,6 +77,7 @@ import java.util.*;
  *
  */
 public class SparkDataSetProcessor implements DistributedDataSetProcessor, Serializable {
+    private static final long serialVersionUID = 9152794997108375878L;
     private long failBadRecordCount = -1;
     private boolean permissive;
     private String statusDirectory;
@@ -97,8 +99,10 @@ public class SparkDataSetProcessor implements DistributedDataSetProcessor, Seria
     // The depth of the current operation being processed via getDataSet
     // in the operation tree.
     private int opDepth = 0;
+    private boolean accumulators;
 
     public SparkDataSetProcessor() {
+        accumulators = EngineDriver.driver().getConfiguration().getSparkAccumulatorsEnabled();
     }
 
     @Override
@@ -162,7 +166,10 @@ public class SparkDataSetProcessor implements DistributedDataSetProcessor, Seria
     @Override
     public <Op extends SpliceOperation> OperationContext<Op> createOperationContext(Op spliceOperation) {
         setupBroadcastedActivation(spliceOperation.getActivation(), spliceOperation);
-        OperationContext<Op> operationContext = new SparkOperationContext<>(spliceOperation, broadcastedActivation);
+        OperationContext<Op> operationContext =
+                accumulators
+                        ? new SparkOperationContext<>(spliceOperation, broadcastedActivation)
+                        : new SparkLeanOperationContext<>(spliceOperation, broadcastedActivation);
         spliceOperation.setOperationContext(operationContext);
         if (permissive) {
             operationContext.setPermissive(statusDirectory, importFileName, failBadRecordCount);
@@ -173,11 +180,10 @@ public class SparkDataSetProcessor implements DistributedDataSetProcessor, Seria
 
     @Override
     public <Op extends SpliceOperation> OperationContext<Op> createOperationContext(Activation activation) {
-        if (activation !=null) {
-            return new SparkOperationContext<>(activation, broadcastedActivation);
-        } else {
-            return new SparkOperationContext<>(activation, null);
-        }
+        BroadcastedActivation ba = activation != null ? broadcastedActivation : null;
+        return accumulators
+                ? new SparkOperationContext<>(activation, ba)
+                : new SparkLeanOperationContext<>(activation, ba);
     }
 
     @Override
@@ -474,7 +480,9 @@ public class SparkDataSetProcessor implements DistributedDataSetProcessor, Seria
                                 .orc(location);
                     } else if (storedAs.toLowerCase().equals("t")) {
                         // spark-2.2.0: commons-lang3-3.3.2 does not support 'XXX' timezone, specify 'ZZ' instead
-                        schema = SpliceSpark.getSession().read().option("timestampFormat", "yyyy-MM-dd'T'HH:mm:ss.SSSZZ").csv(location).schema();
+                        dataset = SpliceSpark.getSession().read().option("timestampFormat", "yyyy-MM-dd'T'HH:mm:ss.SSSZZ").csv(location);
+                    } else {
+                        throw new UnsupportedOperationException("Unsupported storedAs " + storedAs);
                     }
                     dataset.printSchema();
                     schema = dataset.schema();
@@ -793,6 +801,8 @@ public class SparkDataSetProcessor implements DistributedDataSetProcessor, Seria
                     else
                         col=q.negateCompareResult()?col.notEqual(value):col.equalTo(value);
                     break;
+                default:
+                    throw new UnsupportedOperationException("Unknown operator: " + q.getOperator());
             }
             if (andCols ==null)
                 andCols = col;
@@ -827,6 +837,8 @@ public class SparkDataSetProcessor implements DistributedDataSetProcessor, Seria
                     case DataType.ORDER_OP_EQUALS:
                         orCol = q.negateCompareResult() ? orCol.notEqual(value) : orCol.equalTo(value);
                         break;
+                    default:
+                        throw new UnsupportedOperationException("Unknown operator: " + q.getOperator());
                 }
                 if (orCols == null)
                     orCols = orCol;
