+++ splice_access_api/src/main/java/com/splicemachine/access/api/SConfiguration.java
@@ -394,4 +394,5 @@ public interface SConfiguration {
 
     String getRangerUserSyncCaseConversion();
 
+    boolean getSparkAccumulatorsEnabled();
 }
