+++ splice_access_api/src/main/java/com/splicemachine/access/configuration/ConfigurationBuilder.java
@@ -207,6 +207,7 @@ public class ConfigurationBuilder {
     public int regionToLoadPerTask;
     public long controlExecutionRowLimit;
     public int olapShufflePartitions;
+    public boolean sparkAccumulatorsEnabled;
     public String olapLog4jConfig;
     public Map<String,String> olapServerIsolatedRoles;
     public Map<String,String> olapServerYarnQueues;
