+++ splice_access_api/src/main/java/com/splicemachine/access/configuration/OlapConfigurations.java
@@ -117,6 +117,10 @@ public class OlapConfigurations implements ConfigurationDefault {
     public static final String SPARK_COMPACTION_BLOCKING = "spark.compaction.blocking";
     public static final boolean DEFAULT_SPARK_COMPACTION_BLOCKING = true;
 
+    // Whether we record execution stats on Spark accumulators
+    public static final String SPARK_ACCUMULATORS_ENABLED = "spark.accumulators.enabled";
+    public static final boolean DEFAULT_SPARK_ACCUMULATORS_ENABLED = true;
+
     // Log4j config file for OLAP server
     public static final String OLAP_LOG4J_CONFIG = "splice.olap.log4j.configuration";
     public static final String DEFAULT_OLAP_LOG4J_CONFIG = null;
@@ -191,5 +195,6 @@ public class OlapConfigurations implements ConfigurationDefault {
         builder.olapServerIsolatedCompaction = configurationSource.getBoolean(OLAP_SERVER_ISOLATED_COMPACTION, DEFAULT_OLAP_SERVER_ISOLATED_COMPACTION);
         builder.olapServerIsolatedCompactionQueueName = configurationSource.getString(OLAP_SERVER_ISOLATED_COMPACTION_QUEUE_NAME, DEFAULT_OLAP_SERVER_ISOLATED_COMPACTION_QUEUE_NAME);
         builder.olapCompactionAutomaticallyPurgeDeletedRows = configurationSource.getBoolean(OLAP_COMPACTION_AUTOMATICALLY_PURGE_DELETED_ROWS, DEFAULT_OLAP_COMPACTION_AUTOMATICALLY_PURGE_DELETED_ROWS);
+        builder.sparkAccumulatorsEnabled = configurationSource.getBoolean(SPARK_ACCUMULATORS_ENABLED, DEFAULT_SPARK_ACCUMULATORS_ENABLED);
     }
 }
