+++ splice_access_api/src/main/java/com/splicemachine/access/configuration/SConfigurationImpl.java
@@ -161,6 +161,7 @@ public final class SConfigurationImpl implements SConfiguration {
     private final int olapServerMemory;
     private final int olapServerMemoryOverhead;
     private final int olapServerVirtualCores;
+    private final boolean sparkAccumulatorsEnabled;
     private final String olapLog4jConfig;
     private final Map<String, String> olapServerIsolatedRoles;
     private final Map<String, String> olapServerYarnQueues;
@@ -673,6 +674,10 @@ public final class SConfigurationImpl implements SConfiguration {
     public int getOlapShufflePartitions() {
         return olapShufflePartitions;
     }
+    @Override
+    public boolean getSparkAccumulatorsEnabled() {
+        return sparkAccumulatorsEnabled;
+    }
 
     @Override
     public boolean getOlapCompactionAutomaticallyPurgeDeletedRows() {
@@ -1032,6 +1037,7 @@ public final class SConfigurationImpl implements SConfiguration {
         olapServerVirtualCores = builder.olapServerVirtualCores;
         olapClientRetries = builder.olapClientRetries;
         olapShufflePartitions = builder.olapShufflePartitions;
+        sparkAccumulatorsEnabled = builder.sparkAccumulatorsEnabled;
         sparkResultStreamingBatches = builder.sparkResultStreamingBatches;
         sparkResultStreamingBatchSize = builder.sparkResultStreamingBatchSize;
         compactionReservedSlots = builder.compactionReservedSlots;
