+++ splice_machine/src/main/java/com/splicemachine/SpliceKryoRegistry.java
@@ -333,8 +333,7 @@ public class SpliceKryoRegistry implements KryoPool.KryoRegistry{
                 int size=input.readInt();
                 byte[] data=new byte[size];
                 //noinspection ResultOfMethodCallIgnored
-                int bytesRead=input.read(data);
-                assert bytesRead == size: "Did not read entire data point!";
+                input.readBytes(data);
                 dvd.setValue(data);
             }
         },33);
@@ -350,8 +349,7 @@ public class SpliceKryoRegistry implements KryoPool.KryoRegistry{
             protected void readValue(Kryo kryo,Input input,SQLVarbit dvd) throws StandardException{
                 int size=input.readInt();
                 byte[] data=new byte[size];
-                int read=input.read(data);
-                assert read== size: "Did not read entire data";
+                input.readBytes(data);
                 dvd.setValue(data);
             }
         },34);
@@ -367,8 +365,7 @@ public class SpliceKryoRegistry implements KryoPool.KryoRegistry{
             protected void readValue(Kryo kryo,Input input,SQLLongVarbit dvd) throws StandardException{
                 int size=input.readInt();
                 byte[] data=new byte[size];
-                int read=input.read(data);
-                assert read==size: "Did not read entire data";
+                input.readBytes(data);
                 dvd.setValue(data);
             }
         },35);
@@ -618,8 +615,7 @@ public class SpliceKryoRegistry implements KryoPool.KryoRegistry{
             protected void readValue(Kryo kryo,Input input,SQLBlob dvd) throws StandardException{
                 int size=input.readInt();
                 byte[] data=new byte[size];
-                int read = input.read(data);
-                assert read==size: "Did not read entire data line!";
+                input.readBytes(data);
                 dvd.setValue(data);
             }
         },132);
